% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_favicon_design}
\alias{build_favicon_design}
\title{Build the favicon design section of the request}
\usage{
build_favicon_design(
  ios = ios_helper(),
  windows = windows_helper(),
  firefox_app = firefox_app_helper(),
  android_chrome = android_chrome_helper(),
  safari_pinned_tab = safari_pinned_tab_helper(),
  coast = coast_helper(),
  open_graph = open_graph_helper(),
  yandex_browser = yandex_browser_helper()
)
}
\arguments{
\item{ios}{list; list of ios parameters. Most easily created using the \code{ios_helper()} function.}

\item{windows}{list; list of ios parameters. Most easily created using the \code{windows_helper()} function.}

\item{firefox_app}{list; list of ios parameters. Most easily created using the \code{firefox_app_helper()} function.}

\item{android_chrome}{list; list of ios parameters. Most easily created using the \code{android_chrome_helper()} function.}

\item{safari_pinned_tab}{list; list of ios parameters. Most easily created using the \code{safari_pinned_tab_helper()} function.}

\item{coast}{list; list of ios parameters. Most easily created using the \code{coast_helper()} function.}

\item{open_graph}{list; list of ios parameters. Most easily created using the \code{open_graph_helper()} function.}

\item{yandex_browser}{list; list of ios parameters. Most easily created using the \code{yandex_browser_helper()} function.}
}
\value{
list; complete list of favicon parameters for each platform.
}
\description{
This function helps with the creation of the 'favicon_design' section of the request. In turn, this function is fed by
a number of other helper functions designed for each platform.
Further details can be found in the \href{https://realfavicongenerator.net/api/non_interactive_api}{realfavicongenerator.net API documentation}
}
