% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_json_request}
\alias{build_json_request}
\title{Build the complete JSON request}
\usage{
build_json_request(
  api_key,
  master_picture,
  favicon_design,
  settings,
  versioning
)
}
\arguments{
\item{api_key}{character; API key from realfavicongenerator.net}

\item{master_picture}{list; list produced by the \code{build_master_picture()} function that
includes the image or link to image}

\item{favicon_design}{list; embedded list containing the favicon design for different platforms, created using the
\code{build_favicon_design()} function.}

\item{settings}{list; embedded list containing extra settings for request. Created using
the \code{build_settings()} function}

\item{versioning}{list; embedded list of versioning parameters. Created using the \code{build_versioning()} function}
}
\value{
list; list containing the complete request to be converted to JSON and posted
}
\description{
Builds the final JSON request from its constituent parts.
}
