% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_master_picture}
\alias{build_master_picture}
\title{Build the master picture section of the request}
\usage{
build_master_picture(type, value)
}
\arguments{
\item{type}{string; one of "url" for image urls or "local" for local images.}

\item{value}{string; either the url if type == "url" or path to local image}
}
\value{
list; valid list form for the complete JSON request
}
\description{
This helper function creates the 'master_picture' part of the JSON request that includes the image or link to the image
This function is used by the \code{build_json_request} function to create the complete JSON request.
}
