% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_settings}
\alias{build_settings}
\title{Build the settings part of the request}
\usage{
build_settings(
  compression = c("0", "1", "2", "3", "4", "5"),
  scaling_algorithm = c("Mitchell", "NearestNeighbor", "Cubic", "Bilinear", "Lanczos",
    "Spline"),
  error_on_image_too_small = TRUE,
  readme_file = FALSE,
  html_code_file = FALSE,
  use_path_as_is = FALSE
)
}
\arguments{
\item{compression}{character; 1-5 level of compression (5 = higest)}

\item{scaling_algorithm}{character; which scaling algorithm to use. One of "Mitchell", "NearestNeighbor", "Cubic", "Bilinear", "Lanczos", "Spline"}

\item{error_on_image_too_small}{logical; whether to error if an icon cannot be produced because the image is too small}

\item{readme_file}{logical; whether to inclued a README.md file to summarise the package manual installation steps}

\item{html_code_file}{logical; whether to generate a html_code.html file which contains the icons' HTML markups}

\item{use_path_as_is}{logical; FALSE lets RealFaviconGenerator decide the paths for you. TRUE forces RealFaviconGenerator to insert
the path everywhere. Refer to the \href{https://realfavicongenerator.net/api/non_interactive_api}{API documentation} for more.}
}
\description{
This function helps with the creation of the settings list used in the API request.
Further details can be found in the \href{https://realfavicongenerator.net/api/non_interactive_api}{realfavicongenerator.net API documentation}
}
