% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_versioning}
\alias{build_versioning}
\title{Build the versioning part of the request
This function helps with the creation of the versioning list used in the API request.
Further details can be found in the \href{https://realfavicongenerator.net/api/non_interactive_api}{realfavicongenerator.net API documentation}}
\usage{
build_versioning(versioning = FALSE, param_name = NULL, param_value = NULL)
}
\arguments{
\item{versioning}{logical; whether to version the files}

\item{param_name}{character; versioning prefix. Only used if 'versioning' is TRUE}

\item{param_value}{character; version. Only used if 'versioning' is TRUE}
}
\description{
Build the versioning part of the request
This function helps with the creation of the versioning list used in the API request.
Further details can be found in the \href{https://realfavicongenerator.net/api/non_interactive_api}{realfavicongenerator.net API documentation}
}
