% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{firefox_app_helper}
\alias{firefox_app_helper}
\title{Further details can be found in the \href{https://realfavicongenerator.net/api/non_interactive_api}{realfavicongenerator.net API documentation}}
\usage{
firefox_app_helper(
  picture_aspect = c("no_change", "circle", "rounded_square", "square"),
  background_colour = NULL,
  margin = NULL,
  keep_picture_in_circle = TRUE,
  circle_inner_margin = NULL,
  overlay = TRUE,
  manifest = list(app_name = NULL, app_description = NULL, developer_name = NULL,
    developer_url = NULL),
  existing_manifest = NULL,
  on_conflict = c("raise_error", "override", "keep_existing")
)
}
\description{
Further details can be found in the \href{https://realfavicongenerator.net/api/non_interactive_api}{realfavicongenerator.net API documentation}
}
