% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_api_key}
\alias{get_api_key}
\title{Retrieve a previously set API key}
\usage{
get_api_key()
}
\description{
This function returns the value of the API key set at the system environment level.
It will error if the key hasn't been set.
}
\examples{
set_api_key("123abc")
get_api_key()
}
