% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{ios_helper}
\alias{ios_helper}
\title{Create iOS favicon parameters}
\usage{
ios_helper(
  picture_aspect = c("no_change", "background_and_margin"),
  margin = NULL,
  background_colour = NULL,
  startup_image = NULL,
  app_name = NULL,
  assets = list(ios6_and_prior_icons = FALSE, ios7_and_later_icons = TRUE,
    precomposed_icons = FALSE, declare_only_default_icon = TRUE)
)
}
\arguments{
\item{picture_aspect}{character; one of 'no_change' or 'background_and_margin'.
'no_change' leaves the picture as is, 'background_and_margin' uses the background_color and
margin parameters.}

\item{margin}{character; margin of picture in pixels or as a percentage}

\item{startup_image}{list; list containing details of startup image to use (created via \code{build_master_picture}).
Leave as NULL for no startup image}

\item{app_name}{character; when defined, will be used by Safari as the default home screen caption instead of page title}

\item{assets}{list; list of asset parameters. These define which icons to include. Options to be set to TRUE or FALSE
are 'ios6_and_prior_icons', 'ios7_and_later_icons', 'precomposed_icons' and 'declare_only_default_icon'.}

\item{background_color}{character; background color applied to the background of the Apple Touch icon. Hex value.}
}
\description{
This function helps with the creation of the iOS parameter structure in the JSON request.
Further details can be found in the \href{https://realfavicongenerator.net/api/non_interactive_api}{realfavicongenerator.net API documentation}
}
