% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{local_image}
\alias{local_image}
\title{Encode a local favicon image to be passed to the API}
\usage{
local_image(image_path)
}
\arguments{
\item{image_path}{character; path to the image}
}
\value{
list; list containing the embedded image and required extra parameters needed for the JSON request
}
\description{
This function takes a local image path and returns a formatted list expected by the JSON request
}
