% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{open_graph_helper}
\alias{open_graph_helper}
\title{Further details can be found in the \href{https://realfavicongenerator.net/api/non_interactive_api}{realfavicongenerator.net API documentation}}
\usage{
open_graph_helper(
  picture_aspect = c("no_change", "background_and_margin"),
  margin = NULL,
  background_color = NULL,
  ratio = NULL,
  site_url = NULL
)
}
\description{
Further details can be found in the \href{https://realfavicongenerator.net/api/non_interactive_api}{realfavicongenerator.net API documentation}
}
