% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_response}
\alias{parse_response}
\title{Parse an API response}
\usage{
parse_response(response)
}
\arguments{
\item{response}{\code{response()} object}
}
\value{
parsed response
}
\description{
Return the favicon generation result from the \code{response()} object
}
