% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{remove_empty_list}
\alias{remove_empty_list}
\title{Create a list excluding empty elements}
\usage{
remove_empty_list(...)
}
\arguments{
\item{...}{values to be coerced to a list}
}
\value{
list; list without empty values
}
\description{
Create a list excluding empty elements
}
