% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{remove_null_list}
\alias{remove_null_list}
\title{Create a list and remove NULLs}
\usage{
remove_null_list(...)
}
\arguments{
\item{...}{values to be added to list}
}
\value{
list; list without NULL values
}
\description{
Useful to remove NULL values when creating the favicon_design
}
