% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{safari_pinned_tab_helper}
\alias{safari_pinned_tab_helper}
\title{Further details can be found in the [realfavicongenerator.net API documentation](https://realfavicongenerator.net/api/non_interactive_api)}
\usage{
safari_pinned_tab_helper(
  picture_aspect = c("no_change", "silhouette", "black_and_white"),
  theme_color = NULL
)
}
\description{
Further details can be found in the [realfavicongenerator.net API documentation](https://realfavicongenerator.net/api/non_interactive_api)
}
