% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{save_file}
\alias{save_file}
\title{Save the favicon hosted on the realfavicongenerator.net servers}
\usage{
save_file(url, save_loc)
}
\arguments{
\item{url}{character; url of the favicon}

\item{save_loc}{character; directory to save the favicon to}
}
\value{
path to file is returned invisibly
}
\description{
Save the favicon hosted on the realfavicongenerator.net servers
}
