% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send.R
\name{send_request}
\alias{send_request}
\title{Send an API request to realfavicongenerator.net}
\usage{
send_request(json_request)
}
\arguments{
\item{json_request}{list; list to be converted to JSON and sent}
}
\value{
\code{response()} object
}
\description{
Send a preconstructed JSON request (as a list) to the realfavicongenerator.net servers
}
