% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{set_api_key}
\alias{set_api_key}
\title{Set your realfavicongenerator API key globally}
\usage{
set_api_key(api_key)
}
\arguments{
\item{api_key}{character; API key as a character string. Retrieved from realfavicongenerator.net}
}
\description{
By setting your API with this function, you don't need to set it each time you make a request.
The API can be overridden on specific requests by providing a new key when using the \code{generate_favicon()} function.
}
\examples{
set_api_key("1234abcd")
}
