% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{url_image}
\alias{url_image}
\title{Provide an image via url to be passed to the API}
\usage{
url_image(url)
}
\arguments{
\item{url}{character; url to image}
}
\value{
list; list containing the url and required extra parameters needed for the JSON request
}
\description{
This function takes a url to an image and returns a formatted list needed by the JSON request
}
