% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{windows_helper}
\alias{windows_helper}
\title{Create Windows favicon parameters}
\usage{
windows_helper(
  picture_aspect = c("no_change", "white_silhouette"),
  background_colour = NULL,
  assets = list(windows_80_ie_tile = FALSE, windows_10_ie_11_edge_tiles = list(small =
    FALSE, medium = TRUE, big = TRUE, rectange = FALSE)),
  existing_manifest = NULL,
  on_conflict = c("raise_error", "override", "keep_existing"),
  app_name = NULL
)
}
\arguments{
\item{picture_aspect}{character; one of 'no_change' or 'white_silhouette'. 'no_change' leaves master picture as is, and 'white_silhouette'
adds a white silhouette}

\item{assets}{list; embedded list for which icons to generate. List is made up of two entries; 'windows_80_ie_tile' which
must be TRUE or FALSE. Second entry must be a list named 'windows_10_ie_11_edge_tiles' containing TRUE or FALSE values
for each of 'small', 'medium', 'big', 'rectangle'}

\item{existing_manifest}{list; if there is an existing browserconfig.xml file then this can be passed via this parameter.}

\item{on_conflict}{character; must be one of 'raise_error', 'override', or 'keep_existing' that defines what to do if the existing
manifest contains entries that the API creates}

\item{app_name}{character; when provided, this will be the app name used on the home screen caption instead of the page title}

\item{background_color}{character; background color to apply as hex, or NULL for none}
}
\description{
This function helps with the creation of the list structure required to provide Windows-specific favicon parameters
to the API.
Further details can be found in the \href{https://realfavicongenerator.net/api/non_interactive_api}{realfavicongenerator.net API documentation}
}
