% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{yandex_browser_helper}
\alias{yandex_browser_helper}
\title{Further details can be found in the [realfavicongenerator.net API documentation](https://realfavicongenerator.net/api/non_interactive_api)}
\usage{
yandex_browser_helper(
  background_color = NULL,
  manifest = list(show_title = TRUE, version = 1, existing_manigest = NULL,
    error_on_override = FALSE)
)
}
\description{
Further details can be found in the [realfavicongenerator.net API documentation](https://realfavicongenerator.net/api/non_interactive_api)
}
