% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{android_chrome_helper}
\alias{android_chrome_helper}
\title{Further details can be found in the \href{https://realfavicongenerator.net/api/non_interactive_api}{realfavicongenerator.net API documentation}}
\usage{
android_chrome_helper(
  picture_aspect = c("no_change", "background_and_margin", "shadow"),
  manifest = list(name = "", start_url = NULL, display = c("browser", "standalone"),
    orientation = c("portrait", "landscape"), theme_color = NULL, existing_manifest =
    NULL),
  on_conflict = c("raise_error", "override", "keep_existing"),
  assets = list(legacy_icon = FALSE, low_resolution_icons = FALSE)
)
}
\description{
Further details can be found in the \href{https://realfavicongenerator.net/api/non_interactive_api}{realfavicongenerator.net API documentation}
}
