% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{coast_helper}
\alias{coast_helper}
\title{Create Cost favicon parameters
This function helps with the creation of the list structure required to provide Coast-specific favicon parameters
to the API.
Further details can be found in the \href{https://realfavicongenerator.net/api/non_interactive_api}{realfavicongenerator.net API documentation}}
\usage{
coast_helper(
  picture_aspect = c("no_change", "background_and_margin"),
  margin = NULL,
  background_color = NULL
)
}
\arguments{
\item{picture_aspect}{character; one of 'no_change' or 'background_and_margin'. If 'background_and_margin', 'margin' and 'background_color'
parameters can be used}

\item{margin}{character; margin in pixels are as a percentage. Only used if 'picture_aspect' is 'background_and_margin'}

\item{background_color;}{character; hex color string. Only used if 'picture_aspect' is 'background_and_margin'}
}
\description{
Create Cost favicon parameters
This function helps with the creation of the list structure required to provide Coast-specific favicon parameters
to the API.
Further details can be found in the \href{https://realfavicongenerator.net/api/non_interactive_api}{realfavicongenerator.net API documentation}
}
