% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{coast_helper}
\alias{coast_helper}
\title{Further details can be found in the \href{https://realfavicongenerator.net/api/non_interactive_api}{realfavicongenerator.net API documentation}}
\usage{
coast_helper(
  picture_aspect = c("no_change", "background_and_margin"),
  margin = NULL,
  background_color = NULL
)
}
\description{
Further details can be found in the \href{https://realfavicongenerator.net/api/non_interactive_api}{realfavicongenerator.net API documentation}
}
