% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{safari_pinned_tab_helper}
\alias{safari_pinned_tab_helper}
\title{Create Safari pinned tab favicon parameters}
\usage{
safari_pinned_tab_helper(
  picture_aspect = c("no_change", "silhouette", "black_and_white"),
  threshold = NULL,
  theme_color = NULL
)
}
\arguments{
\item{picture_aspect}{character; one of 'no_change', 'silhouette' or 'black_and_white'. If 'black_and_white', use threshold to indicate
how to turn colours to black/white. If 'no_change', image must be a SVG file.}

\item{threshold}{character; value of 0 to 100 which determines how colours are changed to black/white}

\item{theme_color}{character; hex color value of the theme colour - usually the dominant color of the master picture}
}
\description{
This function helps with the creation of the list structure required to provide Safari-specific favicon parameters
to the API.
Further details can be found in the \href{https://realfavicongenerator.net/api/non_interactive_api}{realfavicongenerator.net API documentation}
}
