% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{windows_helper}
\alias{windows_helper}
\title{Further details can be found in the [realfavicongenerator.net API documentation](https://realfavicongenerator.net/api/non_interactive_api)}
\usage{
windows_helper(
  picture_aspect = c("no_change", "white_silhouette"),
  background_colour = list(),
  assets = list(windows_80_ie_tile = FALSE, windows_10_ie_11_edge_tiles = list(small =
    FALSE, medium = TRUE, big = TRUE, rectange = FALSE)),
  existing_manifest = NULL,
  on_conflict = c("raise_error", "override", "keep_existing"),
  app_name = NULL
)
}
\description{
Further details can be found in the [realfavicongenerator.net API documentation](https://realfavicongenerator.net/api/non_interactive_api)
}
